backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
gunicorn>=22.0
# SDK — use local path for development, pip package for production
# Local:  pip install -e /path/to/nzr-autofix-sdks/packages/python
# Published: nzr-autofix>=0.1.0
nzr-autofix>=0.1.0
